Updated UI

+++ src/app/components/add-product/add-product.component.ts
@@ -1,22 +1,24 @@
 import { Component, OnInit } from '@angular/core';
+import { log } from 'console';
 import { ProductService } from 'src/app/services/product.service';
 
 @Component({
   selector: 'app-add-product',
   templateUrl: './add-product.component.html',
-  styleUrls: ['./add-product.component.css']
+  styleUrls: ['./add-product.component.css'],
 })
 export class AddProductComponent implements OnInit {
-
   product = {
     productId: null,
     productName: '',
-    price: 0
+    price: 0,
   };
 
   products: any[] = [];
   message = '';
   editingIndex: number | null = null;
+
+  editMode: boolean = false;
 
   constructor(private productService: ProductService) {}
 
@@ -25,7 +27,7 @@ export class AddProductComponent implements OnInit {
   }
 
   loadProducts() {
-    this.productService.getAllProducts().subscribe(res => {
+    this.productService.getAllProducts().subscribe((res) => {
       this.products = res;
     });
   }
@@ -58,21 +60,43 @@ export class AddProductComponent implements OnInit {
   editProduct(index: number) {
     this.editingIndex = index;
     this.product = { ...this.products[index] };
+    this.editMode = true;
   }
 
   deleteProduct(index: number) {
-    const id = this.products[index].productId;
-    this.productService.deleteProduct(id).subscribe(() => {
-      this.message = 'Product deleted';
-      this.loadProducts();
-    });
+    var confirmation = confirm('Are you sure you want to delete this product?');
+    if (confirmation) {
+      const id = this.products[index].productId;
+      this.productService.deleteProduct(id).subscribe(() => {
+        this.message = 'Product deleted';
+        this.loadProducts();
+      });
+    }
   }
 
   resetForm() {
     this.product = {
       productId: null,
       productName: '',
-      price: 0
+      price: 0,
     };
+  }
+
+  updateProduct() {
+    // This method is now handled in submitProduct
+
+    console.log('Product updated:' + this.product.productId);
+    console.log('editingIndex:' + this.editingIndex);
+
+    this.productService.updateProduct(this.product).subscribe((res) => {
+      console.log(this.product);
+      console.log(res);
+      this.message = 'Product updated successfully';
+      this.editingIndex = null;
+      this.editMode = false;
+      this.resetForm();
+      this.loadProducts();
+      window.location.reload();
+    });
   }
 }
